requirements.txt
Django>=5.1
psycopg2-binary>=2.9
gunicorn
djangorestframework>=3.15